+++ packages/framework/src/Framework/Features/Blogging/Models/PostAuthor.php
@@ -6,19 +6,20 @@ namespace Hyde\Framework\Features\Blogging\Models;
 
 use Hyde\Hyde;
 use Stringable;
-use Hyde\Facades\Author;
+use Illuminate\Support\Str;
 use Hyde\Pages\MarkdownPost;
 use Illuminate\Support\Collection;
 use Hyde\Support\Concerns\Serializable;
 use Hyde\Foundation\Kernel\PageCollection;
 use Hyde\Support\Contracts\SerializableContract;
 
-use function is_string;
 use function array_merge;
 use function array_filter;
 
 /**
  * Object representation of a blog post author for the site.
+ *
+ * @see \Hyde\Facades\Author For the facade to conveniently interact with and create authors.
  */
 class PostAuthor implements Stringable, SerializableContract
 {
@@ -27,7 +28,7 @@ class PostAuthor implements Stringable, SerializableContract
     /**
      * The username of the author.
      *
-     * This is the key used to find authors in the config.
+     * This is the key used to find authors in the config and is taken from that array key.
      */
     public readonly string $username;
 
@@ -59,16 +60,16 @@ class PostAuthor implements Stringable, SerializableContract
     /**
      * The author's social media links/handles.
      *
-     * @var array<string, string
+     * @var ?array<string, string> String-to-string map of social media services to their respective handles.
      *
      * @example ['twitter' => 'mr_hyde'] ($service => $handle)
      */
-    public readonly array $socials;
+    public readonly ?array $socials;
 
     /**
-     * Construct a new Post Author object.
+     * Construct a new Post Author instance with the given data.
      *
-     * If your input is in the form of an array, you may rather want to use the `getOrCreate` method.
+     * If your input is in the form of an array, you may rather want to use the `create` method.
      *
      * @param  string  $username
      * @param  string|null  $name
@@ -77,10 +78,10 @@ class PostAuthor implements Stringable, SerializableContract
      * @param  string|null  $avatar
      * @param  array<string, string>  $socials
      */
-    public function __construct(string $username, ?string $name = null, ?string $website = null, ?string $bio = null, ?string $avatar = null, array $socials = [])
+    public function __construct(string $username, ?string $name = null, ?string $website = null, ?string $bio = null, ?string $avatar = null, ?array $socials = null)
     {
-        $this->username = $username;
-        $this->name = $name ?? $username;
+        $this->username = static::normalizeUsername($username);
+        $this->name = $name ?? static::generateName($username);
         $this->website = $website;
         $this->bio = $bio;
         $this->avatar = $avatar;
@@ -88,25 +89,23 @@ class PostAuthor implements Stringable, SerializableContract
     }
 
     /**
-     * Dynamically get or create an author based on a username string or front matter array.
+     * Create a new Post Author instance from an array of data.
      *
-     * @param  string|array{username?: string, name?: string, website?: string, bio?: string, avatar?: string, socials?: array<string, string>}  $data
+     * If you do not supply a username, the name will be used as the username, or 'Guest' if no name is provided.
+     *
+     * @param  array{username?: string, name?: string, website?: string, bio?: string, avatar?: string, socials?: array<string, string>}  $data
      */
-    public static function getOrCreate(string|array $data): static
+    public static function create(array $data): PostAuthor
     {
-        if (is_string($data)) {
-            return static::get($data);
-        }
-
         return new static(...array_merge([
             'username' => static::findUsernameFromData($data),
         ], $data));
     }
 
-    /** Get an Author from the config, or create it with the username. */
-    public static function get(string $username): static
+    /**  Get a Post Author instance by username, or null if not found. */
+    public static function get(string $username): ?static
     {
-        return static::all()->get($username) ?? Author::create($username);
+        return static::all()->get(static::normalizeUsername($username));
     }
 
     /** @return \Illuminate\Support\Collection<string, \Hyde\Framework\Features\Blogging\Models\PostAuthor> */
@@ -140,6 +139,17 @@ class PostAuthor implements Stringable, SerializableContract
     /** @param array{username?: string, name?: string, website?: string} $data */
     protected static function findUsernameFromData(array $data): string
     {
-        return $data['username'] ?? $data['name'] ?? 'Guest';
+        return static::normalizeUsername($data['username'] ?? $data['name'] ?? 'guest');
+    }
+
+    /** @internal */
+    public static function normalizeUsername(string $username): string
+    {
+        return Str::slug($username, '_');
+    }
+
+    protected static function generateName(string $username): string
+    {
+        return Str::headline($username);
     }
 }