+++ packages/framework/tests/Feature/HydeKernelTest.php
@@ -6,6 +6,7 @@ namespace Hyde\Framework\Testing\Feature;
 
 use Hyde\Facades\Author;
 use Hyde\Facades\Features;
+use InvalidArgumentException;
 use Hyde\Foundation\Facades\Pages;
 use Illuminate\Support\Collection;
 use Hyde\Foundation\Facades\Routes;
@@ -575,21 +576,31 @@ class HydeKernelTest extends TestCase
 
         $this->assertEmpty($kernel->authors()->toArray());
 
-        Config::set('hyde.authors', [Author::create('foo')]);
+        Config::set('hyde.authors', ['foo' => Author::create('foo')]);
 
         $this->assertNotEmpty($kernel->authors()->toArray());
     }
 
-    public function testAuthorsUseTheConfigArrayKey()
+    public function testAuthorsUseTheConfigArrayKeyAsTheUsername()
     {
         Config::set('hyde.authors', ['foo' => Author::create('bar')]);
 
         $this->assertSame([
             'foo' => [
-                'username' => 'bar',
+                'username' => 'foo',
                 'name' => 'bar',
             ],
         ], Hyde::authors()->toArray());
+    }
+
+    public function testAuthorsThrowsExceptionWhenUsernameIsNotSet()
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage('Author username cannot be empty. Did you forget to set the author\'s array key?');
+
+        Config::set('hyde.authors', ['' => Author::create('foo')]);
+
+        Hyde::authors();
     }
 }