+++ packages/framework/tests/Unit/Facades/AuthorTest.php
@@ -15,18 +15,48 @@ class AuthorTest extends TestCase
 {
     public function testCreate()
     {
-        $author = Author::create('john_doe', 'John Doe', 'https://johndoe.com');
+        $author = Author::create('John Doe', 'https://johndoe.com');
 
-        $this->assertSame('john_doe', $author->username);
-        $this->assertSame('John Doe', $author->name);
-        $this->assertSame('https://johndoe.com', $author->website);
+        $this->assertIsArray($author);
+        $this->assertFalse(isset($author['username']));
+        $this->assertSame('John Doe', $author['name']);
+        $this->assertSame('https://johndoe.com', $author['website']);
 
-        $this->assertEquals(Author::create('foo', null, null), Author::create('foo'));
+        $this->assertSame([
+            'name' => 'John Doe',
+            'website' => 'https://johndoe.com',
+            'bio' => null,
+            'avatar' => null,
+            'socials' => null,
+        ], $author);
+
+        $this->assertEquals(Author::create('foo'), Author::create('foo'));
+    }
+
+    public function testCreateWithAllParameters()
+    {
+        $author = Author::create('John Doe', 'https://johndoe.com', 'A cool guy', 'https://johndoe.com/avatar.jpg', [
+            'twitter' => 'johndoe',
+            'github' => 'johndoe',
+        ]);
+
+        $this->assertSame([
+            'name' => 'John Doe',
+            'website' => 'https://johndoe.com',
+            'bio' => 'A cool guy',
+            'avatar' => 'https://johndoe.com/avatar.jpg',
+            'socials' => [
+                'twitter' => 'johndoe',
+                'github' => 'johndoe',
+            ],
+        ], $author);
     }
 
     public function testGet()
     {
         $author = Author::get('mr_hyde');
+
+        $this->assertInstanceOf(PostAuthor::class, $author);
 
         $this->assertSame('mr_hyde', $author->username);
         $this->assertSame('Mr. Hyde', $author->name);